+++ examples/sleuth.py
@@ -25,48 +25,40 @@ logging.getLogger('agent').setLevel(logging.DEBUG)
 log = logging.getLogger('agent')
 # ===================================================================
 
-GAMEDIR = Path(project.DATASETS_DIR_PATH) / Path('game1')
-EXPENSE_DATABASE = GAMEDIR / 'expense_database.db'
-EMPLOYEE_DATABASE = GAMEDIR / 'employee_database.db'
-MARKET_TRENDS_DATABASE = GAMEDIR
-COMPETITOR_ANALYSIS_DATABASE = GAMEDIR / 'competition.db'
-CUSTOMER_FEEDBACK_DATABASE = GAMEDIR / 'customer_feedback.db'
+SIMDIR = Path(project.DATASETS_DIR_PATH) / Path('sim1')
+FINANCE_DATABASE = SIMDIR / 'Company_B_finance.db'
+COMPETITION_DATABSE = SIMDIR / 'Company_B_competition.db'
 DATABASES = {
-    'expense_database': EXPENSE_DATABASE,
-    'employee_database': EMPLOYEE_DATABASE,
-    'competition': COMPETITOR_ANALYSIS_DATABASE,
-    'customer_feedback': CUSTOMER_FEEDBACK_DATABASE,
+    'finance_database': (FINANCE_DATABASE, 'This database describes the company financial operation, including costs and revenues.'),
+    'competitor_database': (COMPETITION_DATABSE, 'This database describes activities of the company\'s competitors who may influence the market.')
 }
-PRIMARY_ASSET = (
-    'Profit and Loss Statement for Tectonic Tech Inc.\n'
-    'Fiscal Year: 1st Jan 2021 - 31st Dec 2021\n'
-    '----------------------------------------------------------------\n'
-    '                        |   Q1  |     Q2   |    Q3   |    Q4   | \n'
-    '----------------------------------------------------------------\n'
-    'Revenue\n'
-    'Product Sales           | 7.0M  |    7.05M | 7.1025M | 7.1576M | \n'
-    '----------------------------------------------------------------\n'
-    'Costs\n'
-    'Cost of Goods Sold      |  2.3M |    2.3M  |   2.22M |   2.22M | \n'
-    '----------------------------------------------------------------\n'
-    'Gross Profit            |  4.7M |    4.75M | 4.8825M | 4.9376M | \n'
-    '----------------------------------------------------------------\n'
-    'Operating Expenses\n'
-    'R&D                     | 0.15M |    0.17M |   0.19M |   0.21M | \n'
-    'Marketing               | 0.10M |    0.10M |   0.12M |   0.13M | \n'
-    'Admin Expenses          | 0.05M |    0.05M |   0.06M |   0.06M | \n'
-    'Employee Compensation   | 4.20M |    4.21M |   4.23M |   4.35M | \n'
-    '----------------------------------------------------------------\n'
-    'Operating Income        |  0.2M |   0.22M |  0.2825M | 0.1876M | \n'
-    '----------------------------------------------------------------\n'
-    'Other Income/Expense    | 0.01M |   0.01M |   -0.02M |  -0.02M | \n'
-    '----------------------------------------------------------------\n'
-    'Net Income Before Taxes | 0.21M  |   0.23M | 0.2625M | 0.1676M | \n'
-    'Taxes                   | 0.033M |  0.037M |  0.041M |   0.05M | \n'
-    '----------------------------------------------------------------\n'
-    'Net Profit              | 0.18M  |  0.193M | 0.2215M | 0.1176M | \n'
-    '----------------------------------------------------------------\n'
-)
+PRIMARY_ASSET = """
+Profit and Loss, Techtonic Tech, Inc
+Jan 1, 2022 through Dec 31, 2022
+All units are in $1000.
+--------------------------------------------------------------------
+                          |      Q1 |      Q2 |      Q3 |      Q4 |
+--------------------------------------------------------------------
+Revenue
+Product sales             |   17791 |   22115 |   17165 |   10804 |
+--------------------------------------------------------------------
+Costs
+Cost of goods sold        |   14820 |   18489 |   14273 |    9047 |
+--------------------------------------------------------------------
+GROSS PROFIT              |    2971 |    3626 |    2892 |    1756 |
+--------------------------------------------------------------------
+Operating Expenses
+HR Expenditure            |    1485 |    1505 |    1499 |    1515 |
+Marketing Expenditure     |      59 |      61 |      60 |      61 |
+Operations Expenditure    |      30 |      29 |      31 |      29 |
+R&D Expenditure           |      60 |      60 |      59 |      60 |
+--------------------------------------------------------------------
+OPERATING INCOME          |    1337 |    1971 |    1243 |      91 |
+--------------------------------------------------------------------
+Taxes                     |     241 |     355 |     224 |      16 |
+--------------------------------------------------------------------
+NET PROFIT                |    1096 |    1616 |    1019 |      75 |
+"""
 
 DEFAULT_QUESTION = 'Does anything look unusual in the report?'
 MAX_LLM_CALLS_PER_INTERACTION = 10
@@ -93,28 +85,34 @@ FUNCTION_DESCRIPTIONS = dispatch.convert_function_descs({
 SYSTEM_MESSAGE = (
     "You are a business analyst for Tectonic Tech, Inc.  You have the ability to run sqlite queries "
     "against the company's databses to collect information for the user.\n\n "
-    "Answer the user's questions as best as you can, and prioritize using data from one of the databases.\n\n"
-    "Before running a query check to see if you already have part or all of the answer from "
-    "your interaction history!\n\n"
+    "Techtonic Tech is a product company selling direct to consumers. You are in a small but stable market "
+    "with only one direct competitor.\n\n"
+    "The values of `expense_category` in the `operating_expenses` table in the `finance_database` include "
+    "\"Marketing Expenditure\", \"HR Expenditure\", \"R&D Expenditure\" and \"Operations Expenditure\".\n\n"
+    "Answer the user's questions as best as you can, and prioritize using data from one or more of the databases."
+    "You do not need permission to run a query against a databse, you can just do it, although you should "
+    "check to see if you already have part or all of the answer from your interaction history!\n\n"
     "If you are unable to answer a question, suggest some alternative questions that you can answer.  Your"
     "goal is to be helpful."
 )
 
 
-def get_schema_description(db_connections) -> str:
+def get_schema_description(db_connections, db_descriptions) -> str:
     """
     Construct a description of the DB schema for the LLM by retrieving the
     CREATE commands used to create the tables.
     """
     schema_description = f"There are {len(db_connections)} databases  Here are their names and schemas:\n"
     for i,db in enumerate(db_connections.keys()):
-        schema_description += f'{i+1}. {db}\n'
+        schema_description += f'Database {i+1} name: {db}\n'
+        schema_description += db_descriptions[db] + '\n'
         cur = db_connections[db].cursor()
         tables = cur.execute("select name from sqlite_master where type='table';").fetchall()
         tables = [col[0] for col in tables]
+        schema_description += f'This database has {len(tables)} tables:\n\n'
         for table in tables:
             schema_description += f'Table `{table}:`\n'
-            schema_description += str(get_table_schema(db_connections[db], table)) + '\n'
+            schema_description += str(get_table_schema(db_connections[db], table)) + '\n\n'
         schema_description += '\n'
 
     return schema_description
@@ -143,13 +141,13 @@ def main():
 
     # Get a model
     model = openai.start_chat('gpt-4-0613')
+    # model = openai.start_chat('gpt-3.5-turbo-0613')
 
     # Connect to the DB and get the table names
     log.debug('Getting the DB information.')
-    db_connections = {}
-    for db in DATABASES:
-        db_connections[db] = sqlite3.connect(DATABASES[db])
-    schema_description = get_schema_description(db_connections)
+    db_connections = {db: sqlite3.connect(DATABASES[db][0]) for db in DATABASES}
+    db_description = {db: DATABASES[db][1] for db in DATABASES}
+    schema_description = get_schema_description(db_connections, db_description)
 
     for _,db in db_connections.items():
         db.close()
@@ -167,7 +165,7 @@ def main():
         # Build the functions that the agent can use
         db_connections = {}
         for db in DATABASES:
-            db_connections[db] = sqlite3.connect(DATABASES[db])
+            db_connections[db] = sqlite3.connect(DATABASES[db][0])
 
         def query_database(database: str, query: str) -> str:
             tool, _ = tools.make_sqlite_query_tool(db_connections[database])
@@ -231,6 +229,10 @@ def main():
 
     with gr.Blocks() as demo:
         gr.Markdown('# Chat About Techtonic Tech, Inc\'s P&L')
+        gr.Markdown(
+            "Techtonic Tech is a product company selling direct to consumers.  "
+            "You are in a small but stable market with only one direct competitor."
+        )
         gr.Markdown('```\n' + PRIMARY_ASSET + '```')
 
         question = gr.Textbox(
@@ -248,12 +250,15 @@ def main():
         chatbot = gr.Chatbot(label='Chatbot message log')
 
         gr.Markdown(
-            'You have access to four databases:\n'
+            'You have access to two databases:\n'
             '\n'
-            '1. Expense database with high-level expenses per category.\n'
-            '2. Employee compensation database with the cost of employee taxes, salaries and bonuses.\n'
-            '3. Competition database with competitor\'s rough financial and market data for the year.\n'
-            '4. Customer feedback database with customer feedback by category.'
+            '1. A finance database with details on revenue, product costs and operating expenses.\n'
+            '2. A competition database that estimates the number of products sold and the marketing spend of your competitor.\n\n'
+            'Example questions:\n'
+            '\n'
+            '1. Is there anything unusual in the report?\n'
+            '2. Why did my revenue decrease in the second half of the year?\n'
+            '3. Is the cost of goods sold a factor in declining revenue?\n'
         )
 
         clear_button.add([question, chatbot])
@@ -276,7 +281,7 @@ def main():
             [status, chatbot]
         ).then(answer_wrapper, chatbot, final_answer)
 
-        clear_button.click(lambda: None, None, [chatbot, status, final_answer], queue=False)
+        clear_button.click(lambda: (None, None, None), outputs=[chatbot, status, final_answer], queue=False)
 
     demo.queue()
     demo.launch()
